services/python-user-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
prometheus-client==0.19.0
nats-py==2.6.0
pydantic[email]==2.5.3

# OpenTelemetry for automatic instrumentation
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-sqlalchemy==0.43b0
opentelemetry-instrumentation-logging==0.43b0
opentelemetry-exporter-otlp==1.22.0
